+++ swiyu-core-business-service-app/src/main/java/ch/admin/bj/swiyu/core/business/modules/trust/api/TrustOnboardingSubmissionDto.java
@@ -20,6 +20,7 @@ public record TrustOnboardingSubmissionDto(
     @NotNull String entityEmail,
     @NotNull AddressDto address,
     @NotNull ContactDto contactPerson,
+    @NotNull TrustOnboardingSubmissionTypeDto type,
     @NotNull Long version,
     @NotNull TrustOnboardingSubmissionStatusDto status,
     @NotNull List<ProofOfPossessionDto> proofOfPossessions,