+++ swiyu-core-business-service-app/src/main/java/ch/admin/bj/swiyu/core/business/modules/trust/api/TrustOnboardingSubmissionTypeDto.java
@@ -1,0 +1,10 @@
+package ch.admin.bj.swiyu.core.business.modules.trust.api;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "Submission type for trust onboarding")
+public enum TrustOnboardingSubmissionTypeDto {
+    REGISTRATION,
+    PROFILE_CHANGE,
+    RENEWAL,
+}