+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.42.24
+
+### Changed
+
+- Update ti-trust-onboarding-information-requested-event to 5.0.0.
+
+## 3.42.23
+
+### Added
+
+- Expose submission type on the trust onboarding submission (always REGISTRATION for now; further types later) (EID-6622)
+
 ## 3.42.22
 
 ### Changed